Moves the parse, reflect and testing packages so that they are internal. This keeps the visible API surface of the package clean.
