+++ internal/reflect/singleton.go
