+++ statement_test.go
@@ -3,8 +3,8 @@ package sqlair
 import (
 	"testing"
 
-	"github.com/canonical/sqlair/parse"
-	sqlairtesting "github.com/canonical/sqlair/testing"
+	"github.com/canonical/sqlair/internal/parse"
+	sqlairtesting "github.com/canonical/sqlair/internal/testing"
 	"github.com/stretchr/testify/assert"
 )
